+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Changed
 
-- Error key now is taken from marketingData message code
+- Get error code from `messages` field on `OrderForm`
 
 ## [0.3.1] - 2019-09-05